refactor: вынес общие базовые механизмы в common
