+++ src/bsl/providers/documentationRender.ts
@@ -1,4 +1,4 @@
-import { Signature, Parameter, Member } from "@/bsl/scope/scope";
+import { Member, Parameter, Signature } from "@/common/scope"
 
 export function signatureLabel(method: Member | string, signature: Signature) {
     const name = (method as Member).name ?? method
@@ -8,6 +8,7 @@ export function signatureLabel(method: Member | string, signature: Signature) {
 export function signatureDocumentation(method: Member, signature: Signature) {
     return signature.description === '' ? method.description : signature.description
 }
+
 export function parameterDocumentation(p: Parameter) {
     if (p.description) {
         return {
