+++ src/bsl/scopeProvider.ts
@@ -1,6 +1,6 @@
 import { editor } from 'monaco-editor-core';
 import { Scope, Member, GlobalScope, TypeDefinition, } from '@/common/scope';
-import { Method } from './Symbols';
+import { Method } from '@/common/codeModel';
 import { Accessible } from '@/bsl/tree-sitter';
 import { isModel } from '@/monaco/utils';
 import { EditorScope } from './scope/editorScope';
