+++ src/common/codeModel/index.ts
@@ -1,0 +1,4 @@
+export * from './symbol'
+export * from './module'
+export * from './method'
+export * from './variable'
