+++ src/common/codeModel/module.ts
@@ -1,0 +1,6 @@
+import { Method, ModuleVariable } from "."
+
+export interface Module {
+    vars: ModuleVariable[]
+    methods: Method[]
+}
