+++ src/common/codeModel/symbol.ts
@@ -1,0 +1,11 @@
+export interface Symbol {
+    name: string
+}
+
+export interface SymbolRange extends Symbol {
+    startLine: number
+    startColumn: number
+    endLine: number
+    endColumn: number
+}
+
