+++ src/common/codeModel/variable.ts
@@ -1,0 +1,9 @@
+import { Symbol } from "./symbol"
+
+export interface Variable extends Symbol {
+    type?: string | Promise<string | undefined>
+}
+
+export interface ModuleVariable extends Variable {
+    isExport: boolean
+}
