+++ src/common/scope/index.ts
@@ -1,4 +1,6 @@
-export * from './members'
-export * from './scope'
-export * from './types'
+export * from './model/members'
+export * from './model/scope'
+export * from './model/types'
 export * from './globalScopeManager'
+export * from './localModuleScope'
+export * from './methodScope'
