+++ src/common/scope/methodScope.ts
@@ -1,0 +1,3 @@
+import { BaseScope } from '@/common/scope'
+
+export class MethodScope extends BaseScope { }
