+++ src/common/scope/model/types.ts
